// src/app/generated-api/model/transaccionAhorroResponseDTO.ts
/**
 * API Banco del Cisne
 * API REST para el sistema bancario del Cisne
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@bancodelcisne.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TransaccionAhorroResponseDTO { 
    idTransaccion?: number;
    idCuenta?: number;
    descripcion?: string;
    monto?: number;
    saldoAnterior?: number;
    saldoNuevo?: number;
    tipoTransaccion?: string;
    metodoPago?: string;
    fechaTransaccion?: Date;
}