// src/app/generated-api/model/cuentaAhorroResponseDTO.ts
/**
 * API Banco del Cisne
 * API REST para el sistema bancario del Cisne
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@bancodelcisne.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CuentaAhorroResponseDTO { 
    idCuenta?: number;
    idCliente?: number;
    nombreCliente?: string;
    numeroCuenta?: string;
    saldoActual?: number;
    tasaInteres?: number;
    tipoCuenta?: string;
    estado?: string;
    fechaApertura?: Date;
    ultimaActualizacion?: Date;
}